+++ foliplus/_cdn.py
@@ -5,12 +5,12 @@ across all controls.
 """
 
 # Fullscreen
-LEAFLET_FULLSCREEN = "3"
+LEAFLET_FULLSCREEN_VERSION = "3"
 
 # HeatmapControl
-H3_JS = "4"
-SIMPLE_STATISTICS = "7"
-CHROMA_JS = "2"
+H3_VERSION = "4"
+SIMPLE_STATISTICS_VERSION = "7"
+CHROMA_VERSION = "2"
 
 # MeasureControl
-GCOORD = "1"
+GCOORD_VERSION = "1"